tree-ssa-reassoc.c (find_operand_rank): Cast pointer via intptr_t to long type.
2009-09-08  Kai Tietz  <kai.tietz@onevision.com>

        * tree-ssa-reassoc.c (find_operand_rank): Cast pointer
        via intptr_t to long type.
        (insert_operand_rank): Cast long type via intptr_t to
        pointer type.
        * genattrtab.c (RTL_HASH): Use intptr_t to cast from
        pointer to scalar.
        * c-pretty-print.c (pp_c_tree_decl_identifier): Cast
        from pointer to unsigned via uintptr_t.

From-SVN: r151515

+++ gcc/genattrtab.c
@@ -319,7 +319,7 @@ static struct attr_hash *attr_hash_table[RTL_HASH_SIZE];
 
 /* Here is how primitive or already-shared RTL's hash
    codes are made.  */
-#define RTL_HASH(RTL) ((long) (RTL) & 0777777)
+#define RTL_HASH(RTL) ((intptr_t) (RTL) & 0777777)
 
 /* Add an entry to the hash table for RTL with hash code HASHCODE.  */
